HWKAGENT-52 - do not allow an app to be deployed unless the server has a "deploy" operation defined. this feature is to allow a user to disable deployments by simply removing the deploy operation metadata from the server resource type (as will be the case for Swarm app server builds)
Notice I moved the "deploy" operation out of the Deployment resource type into the parent WildFly Server type.
"Deploy" operation creates new deployments - so it doesn't make sense to have a "Deploy" operation on a deployment itself
(you don't deploy an existing deployment.) You can REdeploy a deployment, you can REMOVE a deployment, but it is incoherent to say
you are deploying an existing deployment - its already deployed!
You deploy a new deployment on a server, which is why the operation makes more sense being on the server resource type.